+++ pkg/packetbrokeragent/grpc_pba.go
@@ -76,9 +76,15 @@ func (s *pbaServer) GetInfo(ctx context.Context, _ *pbtypes.Empty) (*ttnpb.Packe
 		registration = nil
 	}
 
+	// Register and deregister is only available if Packet Broker Agent is configured with NetID level authorization, and
+	// if the registration is a tenant within that NetID.
+	id, err := s.authenticator.AuthInfo(ctx)
+	registerEnabled := err == nil && id.TenantId == "" && tenantID != ""
+
 	res := &ttnpb.PacketBrokerInfo{
 		ForwarderEnabled:   s.forwarderConfig.Enable,
 		HomeNetworkEnabled: s.homeNetworkConfig.Enable,
+		RegisterEnabled:    registerEnabled,
 	}
 	if registration != nil {
 		res.Registration = &ttnpb.PacketBrokerNetwork{
@@ -101,7 +107,7 @@ var (
 	errRegistration = errors.Define("registration", "get registration information")
 )
 
-func (s *pbaServer) Register(ctx context.Context, _ *pbtypes.Empty) (*ttnpb.PacketBrokerNetwork, error) {
+func (s *pbaServer) Register(ctx context.Context, req *ttnpb.PacketBrokerRegisterRequest) (*ttnpb.PacketBrokerNetwork, error) {
 	if err := rights.RequireIsAdmin(ctx); err != nil {
 		return nil, err
 	}
@@ -127,6 +133,7 @@ func (s *pbaServer) Register(ctx context.Context, _ *pbtypes.Empty) (*ttnpb.Pack
 	if err != nil {
 		return nil, errRegistration.WithCause(err)
 	}
+	listed := req.Listed != nil && req.Listed.Value || req.Listed == nil && registration.Listed
 	devAddrBlocks := toPBDevAddrBlocks(registration.DevAddrBlocks)
 	adminContact, technicalContact := toPBContactInfo(registration.ContactInfo)
 
@@ -139,7 +146,7 @@ func (s *pbaServer) Register(ctx context.Context, _ *pbtypes.Empty) (*ttnpb.Pack
 				DevAddrBlocks:         devAddrBlocks,
 				AdministrativeContact: adminContact,
 				TechnicalContact:      technicalContact,
-				Listed:                registration.Listed,
+				Listed:                listed,
 			},
 		})
 	} else {
@@ -159,7 +166,7 @@ func (s *pbaServer) Register(ctx context.Context, _ *pbtypes.Empty) (*ttnpb.Pack
 				Value: technicalContact,
 			},
 			Listed: &pbtypes.BoolValue{
-				Value: registration.Listed,
+				Value: listed,
 			},
 		})
 	}
@@ -176,7 +183,7 @@ func (s *pbaServer) Register(ctx context.Context, _ *pbtypes.Empty) (*ttnpb.Pack
 		Name:          registration.Name,
 		DevAddrBlocks: fromPBDevAddrBlocks(devAddrBlocks),
 		ContactInfo:   fromPBContactInfo(adminContact, technicalContact),
-		Listed:        registration.Listed,
+		Listed:        listed,
 	}, nil
 }