+++ pkg/packetbrokeragent/mock/pbmapper.go
@@ -16,15 +16,12 @@ package mock
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"testing"
 
 	pbtypes "github.com/gogo/protobuf/types"
 	mappingpb "go.packetbroker.org/api/mapping/v2"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
 )
 
 // PBMapper is a mock Packet Broker Mapper.
@@ -34,15 +31,9 @@ type PBMapper struct {
 }
 
 // NewPBMapper instantiates a new mock Packet Broker Data Plane.
-func NewPBMapper(tb testing.TB, cert tls.Certificate, clientCAs *x509.CertPool) *PBMapper {
-	creds := credentials.NewTLS(&tls.Config{
-		Certificates: []tls.Certificate{cert},
-		ClientAuth:   tls.RequireAndVerifyClientCert,
-		ClientCAs:    clientCAs,
-	})
+func NewPBMapper(tb testing.TB) *PBMapper {
 	mp := &PBMapper{
 		Server: grpc.NewServer(
-			grpc.Creds(creds),
 			grpc.UnaryInterceptor(func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (resp interface{}, err error) {
 				ctx = test.ContextWithTB(ctx, tb)
 				return handler(ctx, req)