+++ pkg/ttnpb/packetbrokeragent.pb.setters.fm.go
@@ -409,6 +409,36 @@ func (dst *PacketBrokerInfo) SetFields(src *PacketBrokerInfo, paths ...string) e
 				var zero bool
 				dst.HomeNetworkEnabled = zero
 			}
+		case "register_enabled":
+			if len(subs) > 0 {
+				return fmt.Errorf("'register_enabled' has no subfields, but %s were specified", subs)
+			}
+			if src != nil {
+				dst.RegisterEnabled = src.RegisterEnabled
+			} else {
+				var zero bool
+				dst.RegisterEnabled = zero
+			}
+
+		default:
+			return fmt.Errorf("invalid field: '%s'", name)
+		}
+	}
+	return nil
+}
+
+func (dst *PacketBrokerRegisterRequest) SetFields(src *PacketBrokerRegisterRequest, paths ...string) error {
+	for name, subs := range _processPaths(paths) {
+		switch name {
+		case "listed":
+			if len(subs) > 0 {
+				return fmt.Errorf("'listed' has no subfields, but %s were specified", subs)
+			}
+			if src != nil {
+				dst.Listed = src.Listed
+			} else {
+				dst.Listed = nil
+			}
 
 		default:
 			return fmt.Errorf("invalid field: '%s'", name)