+++ pkg/webui/console/store/reducers/packet-broker.js
@@ -39,7 +39,9 @@ import {
 const defaultState = {
   info: {},
   registered: false,
+  registerEnabled: false,
   enabled: false,
+  listed: false,
   defaultHomeNetworkRoutingPolicy: {},
   networks: {
     entities: {},
@@ -67,11 +69,14 @@ export default handleActions(
       ...state,
       info: payload,
       registered: Boolean(payload.registration),
+      registerEnabled: Boolean(payload.register_enabled),
+      listed: payload.registration && Boolean(payload.registration.listed),
       enabled: true,
     }),
-    [REGISTER_PACKET_BROKER_SUCCESS]: state => ({
+    [REGISTER_PACKET_BROKER_SUCCESS]: (state, { payload }) => ({
       ...state,
       registered: true,
+      listed: Boolean(payload.listed),
     }),
     [DEREGISTER_PACKET_BROKER_SUCCESS]: state => ({
       ...state,